refactor: svg 컴포넌트 .tsx -> .svg 파일로 변경
