+++ src/assets/index.ts
@@ -6,3 +6,11 @@ export { default as Like } from './svg/like.svg';
 export { default as Share } from './svg/share.svg';
 export { default as Star } from './svg/star.svg';
 export { default as Hearts } from './svg/Hearts.svg';
+export { default as Email } from './svg/Email.svg';
+export { default as Home } from './svg/Home.svg';
+export { default as Key } from './svg/Key.svg';
+export { default as Lock } from './svg/Lock.svg';
+export { default as LeftButton } from './svg/LeftButton.svg';
+export { default as RightButton } from './svg/RightButton.svg';
+export { default as Person } from './svg/Person.svg';
+export { default as Rank } from './svg/Rank.svg';
