--- src/assets/svg/Home.tsx
@@ -1,17 +1,0 @@
-import React from 'react';
-
-const Home = () => {
-  return (
-    <svg
-      width="24"
-      height="21"
-      viewBox="0 0 24 21"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path d="M9.5 19.7499V13.4999H14.5V19.7499C14.5 20.4374 15.0625 20.9999 15.75 20.9999H19.5C20.1875 20.9999 20.75 20.4374 20.75 19.7499V10.9999H22.875C23.45 10.9999 23.725 10.2874 23.2875 9.9124L12.8375 0.499902C12.3625 0.0749023 11.6375 0.0749023 11.1625 0.499902L0.712501 9.9124C0.287501 10.2874 0.55 10.9999 1.125 10.9999H3.25V19.7499C3.25 20.4374 3.8125 20.9999 4.5 20.9999H8.25C8.9375 20.9999 9.5 20.4374 9.5 19.7499Z" />
-    </svg>
-  );
-};
-
-export default Home;
