--- src/assets/svg/LeftButton.tsx
@@ -1,30 +1,0 @@
-import React from 'react';
-
-const LeftButton = () => {
-  return (
-    <svg
-      width="12"
-      height="12"
-      viewBox="0 0 12 12"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g clipPath="url(#clip0_625_2161)">
-        <rect width="12" height="12" fill="white" fillOpacity="0.01" />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M8.51519 1.23371C8.55012 1.26854 8.57782 1.30992 8.59673 1.35548C8.61563 1.40104 8.62536 1.44988 8.62536 1.49921C8.62536 1.54853 8.61563 1.59737 8.59673 1.64293C8.57782 1.68849 8.55012 1.72987 8.51519 1.76471L4.27994 5.99921L8.51519 10.2337C8.58561 10.3041 8.62517 10.3996 8.62517 10.4992C8.62517 10.5988 8.58561 10.6943 8.51519 10.7647C8.44478 10.8351 8.34928 10.8747 8.24969 10.8747C8.15011 10.8747 8.05461 10.8351 7.98419 10.7647L3.48419 6.26471C3.44927 6.22987 3.42156 6.18849 3.40266 6.14293C3.38375 6.09737 3.37402 6.04853 3.37402 5.99921C3.37402 5.94988 3.38375 5.90104 3.40266 5.85548C3.42156 5.80992 3.44927 5.76854 3.48419 5.73371L7.98419 1.23371C8.01903 1.19878 8.06041 1.17108 8.10597 1.15217C8.15153 1.13327 8.20037 1.12354 8.24969 1.12354C8.29902 1.12354 8.34786 1.13327 8.39342 1.15217C8.43898 1.17108 8.48036 1.19878 8.51519 1.23371Z"
-          fill="#393E46"
-        />
-      </g>
-      <defs>
-        <clipPath id="clip0_625_2161">
-          <rect width="12" height="12" fill="white" />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-};
-
-export default LeftButton;
