+++ src/components/common/Profile/Profile.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import Rank from '../../../assets/svg/Rank';
 import ProfileImage from './ProfileImage/ProfileImage';
 
 import {
@@ -9,6 +8,7 @@ import {
 } from './Profile.style';
 import defaultProfile from '../../../assets/images/defaultProfile.png';
 import ProfileInfo from './ProfileInfo';
+import { Rank } from '../../../assets';
 
 const Profile = () => {
   return (
