+++ src/pages/LoginPage/sections/LoginForm/LoginForm.tsx
@@ -1,11 +1,10 @@
 import React, { useRef } from 'react';
 import { css } from '@emotion/react';
-import Email from '../../../../assets/svg/Email';
-import Key from '../../../../assets/svg/Key';
 import Input from '../../../../components/design/Input/Input';
 import { useLoginForm } from '../../../../hooks/login/useLoginForm';
 import Button from '../../../../components/design/Button/Button';
 import { btnLogin, inputContainer } from './LoginForm.style';
+import { Email, Key } from '../../../../assets';
 
 const LoginForm = () => {
   const inputEmailRef = useRef<HTMLInputElement | null>(null);
